Remove unused benchmark and test scripts
Unused benchmark scripts and test files have been removed from the repository. This cleanup helps maintain the focus on currently relevant scripts and avoids potential confusion from the presence of obsolete or irrelevant code.
