+++ python/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "infinity_sdk"
-version = "0.1.0-dev2"
+version = "0.1.0-dev3"
 dependencies = [
     "sqlglot==11.7.1",
     "pydantic",
