+++ python/test/test_table.py
@@ -21,7 +21,7 @@ class TestTable:
     def test_version(self):
         print(infinity.__version__)
 
-    def test_infinity_thrift(self):
+    def test_table(self):
         """
         target: test table apis
         method:
